% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balamuta_package.R
\docType{package}
\name{balamuta-package}
\alias{balamuta}
\alias{balamuta-package}
\title{balamuta: Balamuta R Package}
\description{
Contained within this package are sets of common functions used
for manipulating colors, graphings, modeling, formating, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/balamuta}
  \item Report bugs at \url{https://github.com/coatless/balamuta/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

}
