% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmisc-package.R
\docType{package}
\name{bmisc-package}
\alias{bmisc}
\alias{bmisc-package}
\title{bmisc: Balamuta Miscellaneous}
\description{
Sets of common functions used for manipulating colors,
detecting and interacting with RStudio, modeling, formating, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/bmisc}
  \item Report bugs at \url{https://github.com/coatless/bmisc/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

}
