% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Celsius to Fahrenheit Conversion}
\usage{
celsius_to_fahrenheit(t_celsius)
}
\arguments{
\item{t_celsius}{Temperature recorded in Celsius.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Celsius to Fahrenheit.
}
