% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{Celsius to Kelvin Conversion}
\usage{
celsius_to_kelvin(t_celsius)
}
\arguments{
\item{t_celsius}{Temperature recorded in Celsius.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Celsius to Kelvin.
}
