% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{char_at}
\alias{char_at}
\title{Character at Position i}
\usage{
char_at(s, i)
}
\arguments{
\item{s}{A \code{string} to extract position from.}

\item{i}{A \code{int} between 1 and length.}
}
\value{
A \code{string} of length 1.
}
\description{
Returns the character at location i inside the string.
}
\examples{

s = "statistics"

# Single
char_at(s, 1)

# Vectorized position
char_at(s, c(2, 3))
}
\author{
James J Balamuta
}
