% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{charAt}
\alias{charAt}
\title{Character at Position i}
\usage{
charAt(s, i)
}
\arguments{
\item{s}{A \code{string} to extract position from.}

\item{i}{A \code{int} between 1 and length.}
}
\value{
A \code{string} of length 1.
}
\description{
Returns the character at position i.
}
\examples{

s = "statistics"
charAt(s,1)
}
\author{
James J Balamuta
}

