% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_autostart}
\alias{dropbox_autostart}
\title{Enable or Disable the autostartup feature}
\usage{
dropbox_autostart(autostart = T, dbpy_loc = "~/")
}
\arguments{
\item{autostart}{A \code{boolean} indicating \code{TRUE} (enabled)
or \code{FALSE} (disabled) autostart of the dropbox daemon.}

\item{dbpy_loc}{A \code{string} indicating the location of
the \code{"dropbox.py"} CLI interface.}
}
\description{
Uses the
}
\details{
Attempts to install Dropbox CLI support
}
\examples{
\dontrun{
# Enable Autostart
dropbox_autostart(T)

# Disable Autostart
dropbox_autostart(F)
}
}

