% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_puburl}
\alias{dropbox_puburl}
\title{Obtain the File URL in a Public Folder}
\usage{
dropbox_puburl(fp, dbpy_loc = "~/")
}
\arguments{
\item{fp}{A \code{string} containing the file path that is within the \code{"Public"} folder.}

\item{dbpy_loc}{A \code{string} indicating the location of
the \code{"dropbox.py"} CLI interface.}
}
\description{
Returns the Public File URL in Dropbox's public folder
}
\author{
James Joseph Balamuta
\dontrun{
# Return the file
dropbox_puburl("path/to/file.R")
}
}

