% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_puburl_live}
\alias{dropbox_puburl_live}
\title{Enable Live View of Dropbox File in Browser}
\usage{
dropbox_puburl_live(fp, dbpy_loc = "~/")
}
\arguments{
\item{fp}{A \code{string} containing the file path that is within the \code{"Public"} folder.}

\item{dbpy_loc}{A \code{string} indicating the location of
the \code{"dropbox.py"} CLI interface.}
}
\description{
Enable Live View of Dropbox File in Browser
}
\details{
To get the "live" aspect of the code view,
we opt to set the shared link to have a \code{"?dl=0"}.
This forces the file to only show in the browser.
For more details please see:
\href{https://www.dropbox.com/en/help/201}{How do I force a file or folder to download or render on dropbox.com?}

This was inspired by \href{http://michaellevy.name/blog/useR-talk-on-teaching-R/}{Michael Levy's UseR2016 Talk}
 (\href{https://channel9.msdn.com/Events/useR-international-R-User-conference/useR2016/Teaching-R-to-200-people-in-a-week#time=07m20s}{See the part at 7:20})
}
\examples{
\dontrun{
# Return the file with live streaming enabled
dropbox_puburl_live("path/to/file.R")
}
}
\author{
James Joseph Balamuta
}

