% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_start}
\alias{dropbox_start}
\title{Start the Dropbox Daemon}
\usage{
dropbox_start(dbpy_loc = "~/")
}
\arguments{
\item{dbpy_loc}{A \code{string} indicating the location of
the \code{"dropbox.py"} CLI interface.}
}
\description{
Turns the dropbox daemon online
}
\examples{
\dontrun{
# Start the daemon
dropbox_start()
}
}
\author{
James Joseph Balamuta
}

