% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_stop}
\alias{dropbox_stop}
\title{Stop the Dropbox Daemon}
\usage{
dropbox_stop(dbpy_loc = "~/")
}
\arguments{
\item{dbpy_loc}{A \code{string} indicating the location of
the \code{"dropbox.py"} CLI interface.}
}
\description{
Turns the dropbox daemon off
}
\examples{
\dontrun{
# Stop the daemon
dropbox_stop()
}
}
\author{
James Joseph Balamuta
}

