% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Fahrenheit to Celsius Conversion}
\usage{
fahrenheit_to_celsius(t_fahrenheit)
}
\arguments{
\item{t_fahrenheit}{Temperature recorded in Fahrenheit.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Fahrenheit to Celsius.
}
\examples{

fahrenheit_to_celsius(92)

fahrenheit_to_celsius(32)
}
