% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{fahrenheit_to_kelvin}
\alias{fahrenheit_to_kelvin}
\title{Fahrenheit to Kelvin to Conversion}
\usage{
fahrenheit_to_kelvin(t_fahrenheit)
}
\arguments{
\item{t_fahrenheit}{Temperature recorded in Fahrenheit.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Fahrenheit to Kelvin.
}
