% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{feature_descale}
\alias{feature_descale}
\title{Remove Feature Scaling of Values}
\usage{
feature_descale(x_norm, x_min, x_max)
}
\arguments{
\item{x_norm}{Normalized Values}

\item{x_min}{Minimum non-normalized value}

\item{x_max}{Maximum non-normalized value}
}
\value{
A \code{numeric} vector.
}
\description{
Undo the normalization of values done with
feature scaling.
}
\details{
Converts from the feature scaled or normalized value given by:

\deqn{x[norm] = (x-x[min])/(x[max]-x[min])}{x_{norm} = \frac{(x-x_{min})}{(x_{max}-x_{min})}}

To the original or denormalized value given by:

\deqn{x = x[norm] * (x[max] - x[min]) + x[min]}{x = x_{norm} * (x_{max} - x_{min}) + x_{min}}
}
\examples{
temperatures = c(94.2, 88.1, 32, 0)

temperatures_norm   = feature_scale(temperatures, min(temperatures), max(temperatures))
temperatures_denorm = feature_descale(temperatures_norm, min(temperatures), max(temperatures))

all.equal(temperatures_denorm, temperatures)
}
