% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{feature_scale}
\alias{feature_scale}
\title{Feature Scale Values}
\usage{
feature_scale(x, x_min, x_max)
}
\arguments{
\item{x}{Values}

\item{x_min}{Minimum non-normalized Values}

\item{x_max}{Maximum non-normalized Values}
}
\value{
A \code{numeric} vector.
}
\description{
Apply a feature scaling normalization procedure on Values
}
\details{
Feature scale equation to normalize values:

\deqn{x[norm] = (x-x[min])/(x[max]-x[min])}{x_{norm} = \frac{(x-x_{min})}{(x_{max}-x_{min})}}
}
\examples{
temperatures = c(94.2, 88.1, 32, 0)

temperatures_norm = feature_scale(temperatures, min(temperatures), max(temperatures))
}
