% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{floor_and_cap}
\alias{floor_and_cap}
\title{Floor and Cap a dataset}
\usage{
floor_and_cap(x)
}
\arguments{
\item{x}{A \code{vector} that has length \eqn{N}.}
}
\value{
A \code{vector} with the values floored and capped.
}
\description{
Floors and Caps a numeric variable.
}
\examples{

# One case version
n = 100

x = rnorm(n)

x[n-1] = -99999
x[n] = 10000

y = floor_and_cap(x)

# Dataset example

d = data.frame(x,y=rnorm(n))

o = sapply(d,floor_and_cap)
}
\author{
James J Balamuta
}

