% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_graphic_driver}
\alias{get_graphic_driver}
\title{Natural Graphics Driver for Operating System}
\usage{
get_graphic_driver()
}
\value{
A \code{string} that is either:
\describe{
\item{\code{"quartz"}}{if on MacOS}
\item{\code{"windows"}}{if on Windows}
\item{\code{"x11"}}{if on Linux or Solaris}
}
}
\description{
Provides the default operating system graphics utility
}
\examples{
# Returns a string depending on test platform
get_graphic_driver()
}
\seealso{
\code{\link{is.rstudio}}
}
\author{
James Joseph Balamuta
}
