% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{gg_color}
\alias{gg_color}
\title{Emulate ggplot2 default color palette}
\usage{
gg_color(n)
}
\arguments{
\item{n}{An \code{integer} indicating how many colors user wants.}
}
\value{
A \code{vector} containing \code{n} colors
}
\description{
Autogenerate a colors according to the ggplot selection mechanism.
}
\details{
This implementation is based off an answer given on StackOverflow by
John Colby.
}
\examples{

# Only 1 color
gg_color(1)

# Five colors
gg_color(5)
}
\references{
https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
\author{
John Colby
}
