% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{install_older_packages}
\alias{install_older_packages}
\title{Install an older package from source}
\usage{
install_older_packages(pkgs, versions, lib, repos = getOption("repos"), ...)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose older versions
should be downloaded from the repositories.}

\item{versions}{character vector of the respective package versions to
download.}

\item{lib}{character vector giving the library directories where to
install the packages. Recycled as needed. If missing, defaults
to the first element of \code{\link[base:libPaths]{base::.libPaths()}}.}

\item{repos}{character vector, the base URL(s) of the repositories to use,
e.g., the URL of a CRAN mirror such as "\verb{https://cloud.r-project.org}".
For more details on supported URL schemes see url.}

\item{...}{Additional parameters passed onto \code{\link[utils:install.packages]{utils::install.packages()}}.}
}
\description{
A light wrapper around \code{\link[utils:install.packages]{utils::install.packages()}} that allows for
direct downloading of an older package's source from the CRAN archive.
}
\references{
Documentation includes parts of the \code{\link[utils:install.packages]{utils::install.packages()}} help
manual.
}
