% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{int_to_hex}
\alias{int_to_hex}
\title{Convert 0-255 to a Hex number}
\usage{
int_to_hex(n)
}
\arguments{
\item{n}{An \code{int}}
}
\value{
A \code{string} of length 2.
}
\description{
This is a helper function for \code{\link{rgb_to_hex}}. This function takes a
single R, G, or B numeric value and converts it to hex.
}
\examples{
int_to_hex(22)
}
