% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_os.R
\name{is.windows}
\alias{is.linux}
\alias{is.macos}
\alias{is.sun}
\alias{is.windows}
\title{Check for an Operating System}
\usage{
is.windows()

is.macos()

is.linux()

is.sun()
}
\value{
Either \code{TRUE} or \code{FALSE}
}
\description{
Performs a check to determine the OS
}
\author{
James Joseph Balamuta
}

