% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ints.R, R/utilities.R
\name{is.whole}
\alias{is.whole}
\title{Integer Check}
\usage{
is.whole(x)

is.whole(x)
}
\arguments{
\item{x}{A \code{numeric} value to check to see if it is an integer.}

\item{x}{A \code{numeric} value to check to see if it is an integer.}
}
\value{
A \code{boolean} value indicating whether the value is an integer or not.

A \code{boolean} value indicating whether the value is an integer or not.
}
\description{
Checks whether the submitted value is an integer

Checks whether the submitted vector of values is an integer / whole number
}
\examples{
is.whole(2.3)
is.whole(4)
is.whole(c(1,2,3))
is.whole(c(.4,.5,.6))
is.whole(c(7,.8,9))
is.whole(2.3)
is.whole(4)
is.whole(c(1,2,3))
is.whole(c(.4,.5,.6))
is.whole(c(7,.8,9))
}
\author{
James J Balamuta

JJB
}
\keyword{internal}

