% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_os.R
\name{is_windows}
\alias{is_windows}
\alias{is_macos}
\alias{is_linux}
\alias{is_sun}
\title{Check for an Operating System}
\usage{
is_windows()

is_macos()

is_linux()

is_sun()
}
\value{
Either \code{TRUE} or \code{FALSE}
}
\description{
Performs a check to determine the OS
}
\author{
James Joseph Balamuta
}
