% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_rstudio}
\alias{is_rstudio}
\title{Is R Open in RStudio?}
\usage{
is_rstudio()
}
\value{
A \code{logical} value that indicates whether R is open in RStudio.
}
\description{
Detects whether R is open in RStudio.
}
\examples{
is_rstudio()
}
\author{
JJB
}
