% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ints.R
\name{is_whole}
\alias{is_whole}
\title{Integer Check}
\usage{
is_whole(x)
}
\arguments{
\item{x}{A \code{numeric} value to check to see if it is an \code{integer}.}
}
\value{
A \code{boolean} value indicating whether the value is an \code{integer} or not.
}
\description{
Checks whether the submitted value is an \code{integer}
}
\examples{
is_whole(2.3)
is_whole(4)
is_whole(c(1,2,3))
is_whole(c(.4,.5,.6))
is_whole(c(7,.8,9))
}
\author{
James J Balamuta
}
