% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{Kelvin to Celsius Conversion}
\usage{
kelvin_to_celsius(t_kelvin)
}
\arguments{
\item{t_kelvin}{Temperature recorded in Kelvin.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Kelvin to Celsius.
}
