% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{kelvin_to_fahrenheit}
\alias{kelvin_to_fahrenheit}
\title{Kelvin to Fahrenheit Conversion}
\usage{
kelvin_to_fahrenheit(t_kelvin)
}
\arguments{
\item{t_kelvin}{Temperature recorded in Kelvin.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Celsius to Kelvin.
}
\examples{
kelvin_to_fahrenheit(92)

kelvin_to_fahrenheit(32)
}
