% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lagged}
\alias{lagged}
\title{Lag Vector Values}
\usage{
lagged(x, lag = 1)
}
\arguments{
\item{x}{A \code{vec} of data.}

\item{lag}{An \code{integer} value.}
}
\value{
A \code{vector} with lagged values and \code{NA}s.
}
\description{
Provides a lagging mechanism for vector data.
}
\examples{
x = rnorm(10)

lagged(x, 2)
}
\author{
James Balamuta
}
