% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mkdir}
\alias{mkdir}
\title{Make Directory}
\usage{
mkdir(dir)
}
\arguments{
\item{dir}{A \code{string} indicating the directory to make.}
}
\value{
New directory on file system
}
\description{
Create a directory using either a relative path or an absolute path.
}
\examples{
\dontrun{
# Make directory from working directory
mkdir("toad")

# Make directory from absolute path
mkdir("C:/path/to/dir/toad")
}
}

