% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{na_prop_overall}
\alias{na_prop_overall}
\alias{na_prop_by_variable}
\alias{na_prop_by_observation}
\alias{na_count_overall}
\alias{na_count_by_variable}
\alias{na_count_by_observation}
\title{Determine Amount of Missing Values in Data}
\usage{
na_prop_overall(x)

na_prop_by_variable(x)

na_prop_by_observation(x)

na_count_overall(x)

na_count_by_variable(x)

na_count_by_observation(x)
}
\arguments{
\item{x}{A vector of length \eqn{N} or a matrix with dimensions \eqn{N \times P}{N x P}.}
}
\value{
\itemize{
\item Overall: a single numeric value between \verb{[0, 1]} or a count between \verb{[0, N]}.
\item Variable: \eqn{P} different numeric values between \verb{[0, 1]} or counts between \verb{[0, N]}.
\item Observation: \eqn{N} different numeric values between \verb{[0, 1]} or counts between \verb{[0, P]}.
}
}
\description{
Computes the proportion of data that is missing in a given data set.
}
\examples{
# By vector
x = c(1, 2, NA, 4)
na_prop_overall(x)
na_count_overall(x)

# By Data Frame
missing_df = data.frame(
 a = c(1, 2, NA, 4),
 b = c(3, NA, 2, NA)
)

# Proportion
na_prop_overall(missing_df)
na_prop_by_variable(missing_df)
na_prop_by_observation(missing_df)

# Counts
na_count_overall(missing_df)
na_count_by_variable(missing_df)
na_count_by_observation(missing_df)

}
