% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{nb2latex}
\alias{nb2latex}
\title{Export Data in Scientific Notation to LaTeX}
\usage{
nb2latex(x, rounding = 4, align = TRUE)
}
\arguments{
\item{x}{A \code{numeric vector} to be formatted.}

\item{rounding}{A \code{integer} that indicates how many decimal places should exist.}

\item{align}{A \code{bool} that indicates whether phantom padding should be added to force column into alignment.}
}
\value{
A \code{character vector} containing the latex notation.
}
\description{
Takes a vector of observations and converts them to have the
scientific notation in LaTeX. E.g. \code{0.0003 => 3 \cdot 10^{-4}}
}
\examples{

# For reproducibility
set.seed(1337)

x = rnorm(5,.0001,.2)

nb2latex(x)

nb2latex(x, rounding = 5)

nb2latex(x, rounding = 5, align = FALSE)
}
\author{
Stephane Guerrier & James Balamuta
}
