% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{pad_number}
\alias{pad_number}
\title{Pad Numeric Numbers}
\usage{
pad_number(x)
}
\arguments{
\item{x}{A \code{vector}}
}
\value{
A \verb{character vector} that is padded to the length of the
maximum entry.
}
\description{
Add zeros before start of the number
}
\examples{
# Padding applied
pad_number(8:10)

# No padding applied
pad_number(2:3)

# Pads non-negative number with 0.
# This needs to be improved slightly...
pad_number(-1:1)
}
\author{
James Balamuta
}
