% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cast.R
\name{plot.simdf}
\alias{autoplot.simdf}
\alias{plot.simdf}
\title{Plot Simulation Trials}
\usage{
\method{plot}{simdf}(x, ...)

autoplot.simdf(object, type = "line", ...)
}
\arguments{
\item{x, object}{An \code{\link{simdf}} object.}
}
\description{
Constructs a line graph containing different simulations
}
\examples{
# Set Seed
set.seed(5812)

# Generate data
m = matrix(rnorm(10), 2,5)

# Organize data.frame by row
sim = cast_simdf(m)

# Graph Sim
plot(sim)
}

