% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_os.R
\name{require_linux}
\alias{require_linux}
\alias{require_windows}
\alias{require_macos}
\alias{require_sun}
\title{Require a Specific Operating System}
\usage{
require_linux()

require_windows()

require_macos()

require_sun()
}
\description{
Mandates the presence of an operating system
}
\details{
If any of these functions are called
on the wrong operating system. A stop error
is triggered and the function will fail.
}
\author{
James Joseph Balamuta
}
