% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert RGB Value to Hexadecimal}
\usage{
rgb2hex(R, G, B, pound = T)
}
\arguments{
\item{n}{An \code{int}}
}
\value{
A \code{string} of length 2.
}
\description{
This is a helper function for \link{rgb2hex}.
}
\examples{
toHex(22)
}

