% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq_bin}
\alias{seq_bin}
\title{Integer Bins with Overflow}
\usage{
seq_bin(n, bins = 2, from = 1)
}
\arguments{
\item{n}{A \code{integer} that is the total number of objects.}

\item{bins}{A \code{integer} indicating the number of bins.}

\item{from}{A \code{integer} indicating where the sequence should start.}
}
\value{
A \code{vector} containing integer binning.
}
\description{
Creates a vector at equal level with the number of items per bin
with the overflow going to the last bin. That is, the last bin
is filled with less elements than the prior bins.
}
\examples{
# Evenly spaced
a = seq_bin(100, 2)

b = seq_bin(100, 4)

# Oddly spaced
d = seq_bin(101,5)

# First four bins equal, last differs.
diff(d)
}

