% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cast.R
\name{sim_df}
\alias{sim_df}
\title{Cast Simulation Matrix to Data Frame}
\usage{
sim_df(m, wide = TRUE)
}
\arguments{
\item{m}{A \code{matrix}}

\item{wide}{A \code{boolean} indicating if the simulated data is group
by row (\code{TRUE}) or column (\code{FALSE}).}
}
\value{
A \code{data.frame} with three variables:
\itemize{
\item{Round}{Iteration of the Simulation}
\item{Draw}{Draw during the iteration of the simulation}
\item{Value}{Value of the statistic at round and draw}
}
}
\description{
Creates a \code{data.frame} out of a simulation matrix
for use with \code{ggplot2}.
}
\examples{
# Set Seed
set.seed(5812)

# Generate data
m = matrix(rnorm(10), 2,5)

# Organize data.frame by row
sim_df(m)

# Organize by column
sim_df(m, wide = FALSE)
}
\author{
James Balamuta
}
