% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cast.R
\name{study_df}
\alias{study_df}
\title{Cast Simulation Study Matrices to Data.Frame}
\usage{
study_df(..., wide = TRUE, data_names = NULL)
}
\arguments{
\item{...}{A list of \code{matrices}}

\item{wide}{A \code{boolean} indicating if the simulated data is group
by row (\code{TRUE}) or column (\code{FALSE}).}

\item{data_names}{A \code{character vector} containing name of matrix. If
empty it uses the simulation matrix name}
}
\value{
A \code{data.frame} with three variables:
\itemize{
\item{Round}{Iteration of the Simulation}
\item{Draw}{Draw during the iteration of the simulation}
\item{Value}{Value of the statistic at round and draw}
\item{Type}{Study Matrix}
}
}
\description{
Creates a \code{data.frame} out of multiple simulation matrices
for use with \code{ggplot2}.
}
\examples{
# Set Seed
set.seed(5812)

# Generate data
m1 = matrix(rnorm(10), 2, 5)
m2 = matrix(rnorm(10), 2, 5)
m3 = matrix(rnorm(10), 2, 5)

# Organize data.frame by row
study_df(m1, m2, m3)

# Organize by column
study_df(m1, m2, m3, wide = FALSE, 
         data_names = c("Hello", "Goodbye", "Wabbit"))
}
\author{
James Balamuta
}

