% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{system_graphic_driver}
\alias{system_graphic_driver}
\title{Natural Graphics Driver for Operating System}
\usage{
system_graphic_driver()
}
\value{
A \code{string} that is either:
\itemize{
\item \code{"quartz"}: if on MacOS
\item \code{"windows"}: if on Windows
\item \code{"x11"}: if on Linux or Solaris
}
}
\description{
Provides the default operating system graphics utility
}
\examples{
# Returns a string depending on test platform
system_graphic_driver()
}
\seealso{
\link{is_rstudio}
}
\author{
James Balamuta
}
