% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{tint}
\alias{tint}
\title{Tint an RGB value}
\usage{
tint(rgb_value, tint_factor = 0.2)
}
\arguments{
\item{rgb_value}{A \code{vector} with length \eqn{3 \times 1}{3 x 1}.}

\item{tint_factor}{A \code{double} that ranges between \eqn{[0, 1]}.}
}
\value{
A \code{matrix} with dimensions \eqn{3 \times 1}{3 x 1}.
}
\description{
The function tints or lightens an RGB value by adding white to the values.
}
\examples{
tint(c(22, 150, 230), tint_factor = 0.5)
}
