% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{toHex}
\alias{toHex}
\title{Convert 0-255 to a Hex number}
\usage{
toHex(n)
}
\arguments{
\item{n}{An \code{int}}
}
\value{
A \code{string} of length 2.
}
\description{
This is a helper function for \link{rgb2hex}.
This function takes a single R, G, or B numeric value and converts it to hex.
}
\examples{
toHex(22)
}
