% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{tr}
\alias{tr}
\title{Obtain the Trace of a Square Matrix}
\usage{
tr(x)
}
\arguments{
\item{x}{A \code{matrix} that is square e.g. \eqn{N \times N}{N x N}}
}
\value{
A \code{matrix} with circles imprinted within its dimensions.
}
\description{
Calculates and returns the trace of a square matrix.
}
\examples{
# I_2 matrix
tr(diag(2))
}
\author{
James J Balamuta
}
