% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{url_title}
\alias{url_title}
\title{Create a "safe" url title}
\usage{
url_title(st)
}
\arguments{
\item{st}{A \code{string} that needs to be a title in a url}
}
\value{
A \code{string} with the aforementioned modifications.
}
\description{
Takes a string, forces characters to lower case, then removes punctuation and switch spaces to - instead of _
}
\examples{
url_title("My Name is Jaime!")
}
\author{
JJB
}
