% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balamuta_package.R
\docType{package}
\name{balamuta-package}
\alias{balamuta-package}
\title{Balamuta R Package}
\description{
The package contains functions useful functions for doing analysis within R, working with graphics, and strings.
}
\details{
\tabular{ll}{
Package: \tab balamuta\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-12-06\cr
License: \tab MIT + file LICENSE\cr
}
}
\author{
James Balamuta \email{balamut2@illinois.edu}

James Balamuta \email{balamut2@illinois.edu}
}

