% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{char_at}
\alias{char_at}
\title{Character at Position \emph{i}}
\usage{
char_at(x, index)
}
\arguments{
\item{x}{A \verb{character vector} to extract position from.}

\item{index}{An \code{integer} between 1 and length \emph{n}.}
}
\value{
A \verb{character vector} of length \code{index}.
}
\description{
Returns the character at location \emph{i} inside the string.
}
\examples{
# Example string
s = "statistics"

# Single character
char_at(s, 1)

# Vectorized position
char_at(s, c(2, 3))
}
\author{
James J Balamuta
}
