% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{install_dropbox_cli}
\alias{install_dropbox_cli}
\title{Install Dropbox CLI version on Linux}
\usage{
install_dropbox_cli()
}
\description{
Downloads and installs the linux Dropbox CLI version
using advice given \url{https://www.dropbox.com/install?os=lnx}
}
\details{
Attempts to install Dropbox CLI support in the default location of 
\code{"~/Dropbox"}, \code{"~/.dropbox-dist/dropboxd"}, and 
\code{"~/dropbox.py"}.
}
\examples{
\dontrun{
# Install Dropbox CLI to local directory
install_dropbox_cli()
}
}

