% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjb-package.R
\docType{package}
\name{jjb-package}
\alias{jjb}
\alias{jjb-package}
\title{jjb: Balamuta Miscellaneous}
\description{
Set of common functions used for manipulating colors,
    detecting and interacting with 'RStudio', modeling, formatting, determining
    users' operating system, feature scaling, and more!
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

}
