% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{kelvin_to_fahrenheit}
\alias{kelvin_to_fahrenheit}
\title{Celsius to Kelvin Conversion}
\usage{
kelvin_to_fahrenheit(t_kelvin)
}
\arguments{
\item{t_celsius}{Temperature recorded in Celsius.}
}
\value{
A \code{numeric} vector.
}
\description{
Converts temperature recorded in Celsius to Kelvin.
}
