% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mkdir}
\alias{mkdir}
\title{Make Directory}
\usage{
mkdir(dir, r = TRUE)
}
\arguments{
\item{dir}{A \code{string} indicating the directory to make.}

\item{r}{A \code{boolean} that indicates whether the directories should be
made recursively}
}
\value{
New directory on file system
}
\description{
Create a directory using either a relative path or an absolute path.
}
\examples{
\donttest{
# Make directory from working directory
mkdir("toad")

## This assumes the computer is on Windows and the C drive exists.
# Make directory from absolute path
mkdir("C:/path/to/dir/toad")
}
}
\author{
James Balamuta
}
