% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cast.R
\name{plot.sim_df}
\alias{autoplot.sim_df}
\alias{plot.sim_df}
\title{Plot Simulation Trials}
\usage{
\method{plot}{sim_df}(x, ...)

\method{autoplot}{sim_df}(object, ...)
}
\arguments{
\item{x, object}{An \code{\link{sim_df}} object.}

\item{...}{Not used...}
}
\description{
Constructs a line graph containing different simulations
}
\examples{
# Set Seed
set.seed(5812)

# Generate data
m = matrix(rnorm(10), 2,5)

# Organize data.frame by row
sim = sim_df(m)

# Correct ggplot2 usage
autoplot(sim)

# Base R 
plot(sim)
}

