% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert RGB Value to Hexadecimal}
\usage{
rgb2hex(R, G, B, pound = TRUE)
}
\arguments{
\item{R}{A \code{int} that is between 0 and 255 for the Red value.}

\item{G}{A \code{int} that is between 0 and 255 for the Green value.}

\item{B}{A \code{int} that is between 0 and 255 for the Blue value.}

\item{pound}{A \code{bool} that indicates whether a pound sign should be prepended to the hexadecimal.}
}
\value{
A \code{string} containing the hexadecimal informatoin.
}
\description{
This function converts an RGB value to the hexadecimal numbering system.
}
\examples{
# Hexadecimal with pound sign
rgb2hex(255,255,255)

# Heaxadecimal without pound sign
rgb2hex(255,255,255,FALSE)
}
